requirements.txt
colorama
requests
python-dotenv
